+++ src/main/java/com/compi/elitewings/repository/IFlightRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface IFlightRepository extends JpaRepository<Flight, Integer> {
+public interface IFlightRepository extends JpaRepository<Flight, Long> {
     List<Flight> findByPurpose(Flight.Purpose purpose);
 }
